Potential fix for code scanning alert no. 16: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done using the larger integer type, and the result will be correctly handled by the opj_calloc function.

The specific change involves casting sa->block_width to unsigned long before multiplying it by sa->block_height on line 238.

+++ libopenjpeg-2.3.1/src/lib/openjp2/sparse_array.c
@@ -235,7 +235,7 @@ static OPJ_BOOL opj_sparse_array_int32_read_or_write(
             } else {
                 if (src_block == NULL) {
                     src_block = (OPJ_INT32*) opj_calloc(1,
-                                                        sa->block_width * sa->block_height * sizeof(OPJ_INT32));
+                                                        (unsigned long)sa->block_width * sa->block_height * sizeof(OPJ_INT32));
                     if (src_block == NULL) {
                         return OPJ_FALSE;
                     }
